Move chain config wrapper type to params/ and add test
